requirements.txt
# 📦 Core test libraries
pytest>=7.0.0,<8.0.0
pytest-asyncio==0.21.1
pytest-html>=4.1.1
pytest-xdist>=3.7.0
pytest-timeout>=2.1.0
pytest-metadata>=3.1.1
pytest-cov>=4.1.0
pytest-base-url>=2.1.0

# 🌐 Playwright for UI testing
playwright>=1.44.0
pytest-playwright>=0.4.3

# 🔗 HTTP Client for REST API testing
httpx>=0.27.0

# 🔌 gRPC and Protobuf
grpcio>=1.62.0
grpcio-tools>=1.62.0
protobuf>=4.25.0

# 🛠️ Utilities
loguru>=0.7.0
python-dotenv>=1.0.0
faker>=20.1.0
jsonschema>=4.21.0
colorama>=0.4.6
rich>=13.7.0

# 🛠️ Additional libraries
requests>=2.32.4

# 📊 Reporting libraries
allure-pytest>=2.13.2
allure-python-commons==2.13.2

# 🛠️ Additional libraries
factory-boy==3.3.0

# 🛠️ Additional libraries
psutil>=6.1.0
memory-profiler==0.61.0

# 🛠️ Additional libraries
pydantic==2.5.0
pydantic-settings==2.1.0

# 🛠️ Additional libraries
click==8.1.7
tqdm==4.66.1

# 🛠️ Additional libraries
sqlalchemy==2.0.23
alembic==1.13.1

# 🛠️ Additional libraries
docker==6.1.3

# 🛠️ Additional libraries
pre-commit==3.6.0

# Data generation and management
Faker>=37.4.0
PyYAML>=6.0.1
jinja2>=3.1.2

# Performance monitoring
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.1
ruff>=0.1.0
